// src/enums/GeneralEnums.tsx
export enum ELayoutType {
    Widescreen = "WIDESCREEN",
    Desktop = "DESKTOP",
    NarrowDesktop = "NARROWDESKTOP",
    Tablet = "TABLET",
    Mobile = "MOBILE",
}
export enum EModalType {
    Info = "INFO",
    AttachmentViewer = "ATTACHMENT_VIEWER",
    RoomEditor = "ROOM_EDITOR",
    Verify = "VERIFY"
}